refactor: fix function structure
