+++ src/components/dialog/form/CheckBoxColumnForm/CheckBoxColumnForm.tsx
@@ -14,7 +14,7 @@ import {
 import {
   InlineDialogHeader,
   InlineDialogTitle,
-} from "@/components/ui/inline-dialog";
+} from "@/components/custom-ui/InlineDialog";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
