+++ src/components/dialog/form/OpenSatisfactionSurveyForm/OpenSatisfactionSurveyForm.tsx
@@ -18,7 +18,7 @@ import {
   InlineDialogFooter,
   InlineDialogHeader,
   InlineDialogTitle,
-} from "@/components/ui/inline-dialog";
+} from "@/components/custom-ui/InlineDialog";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { RadioGroup } from "@/components/ui/radio-group";
