+++ src/components/dialog/form/SignInForm/SignInForm.tsx
@@ -17,7 +17,7 @@ import {
   InlineDialogDescription,
   InlineDialogHeader,
   InlineDialogTitle,
-} from "@/components/ui/inline-dialog";
+} from "@/components/custom-ui/InlineDialog";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { zodResolver } from "@hookform/resolvers/zod";
