+++ src/components/dialog/information/CreditsDialog/CreditsDialog.tsx
@@ -15,7 +15,7 @@ import {
   InlineDialogDescription,
   InlineDialogHeader,
   InlineDialogTitle,
-} from "@/components/ui/inline-dialog";
+} from "@/components/custom-ui/InlineDialog";
 import { Input } from "@/components/ui/input";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { useToast } from "@/hooks/use-toast";
