+++ src/components/dialog/interrupt/UpdateAlert/UpdateAlert.tsx
@@ -15,7 +15,7 @@ import {
   InlineDialogFooter,
   InlineDialogHeader,
   InlineDialogTitle,
-} from "@/components/ui/inline-dialog";
+} from "@/components/custom-ui/InlineDialog";
 import { DownloadCloudIcon } from "lucide-react";
 
 interface UpdateAlertProps {
