+++ src/hooks/useDialogComponent.ts
@@ -11,7 +11,7 @@ import {
   InlineDialogHeader,
   type InlineDialogProps,
   InlineDialogTitle,
-} from "@/components/ui/inline-dialog";
+} from "@/components/custom-ui/InlineDialog";
 import {
   AlertDialog,
   AlertDialogAction,
