+++ src/layouts/Layout.tsx
@@ -1,6 +1,6 @@
 import { Toaster } from "@/components/ui/toaster";
-import { Header } from "@/layouts/Header/Header";
 import { Outlet } from "react-router-dom";
+import Header from "@/layouts/Header/Header";
 
 export default function Layout() {
   return (
